requirements.txt
flask
numpy
pillow
tensorflow-cpu
